changes
started with experimenting with the space complexity

+++ Assets/AStar/Scripts/Core/Data/EvaluationData.cs
@@ -3,8 +3,8 @@
 [Serializable]
 public class EvaluationData
 {
-    public float TimeTaken;
-    public float SpaceTaken;
+    public double TimeTaken;
+    public long SpaceTaken;
     public int PathLength;
     public float PathCost;
     public int CorridorIterations;
